fix: Use network name in switch alert
In the send flow, scanning a QR code to fill out the "to" field can
also trigger a network switch, which shows an alert to the user.
Elsewhere in the codebase where this alert is shown, the network name
is used in the alert. But here in the send flow, the network ID was
used.

Now the network name is used everywhere. The network name is the name
users are likely to understand; we don't expect users to memorize
or understand network IDs.

This relates to https://github.com/MetaMask/mobile-planning/issues/1226