fix: Fix incoming transaction notifications on built-in networks
Incoming transaction notifications were not working correctly on built-
in networks. This was due to a filter intended to ensure notifications
were only emitted for the current selected network. This filter used
the `tx.transaction.networkID` to match the transaction against the
current network, but this property does not exist for incoming
transactions. Instead the network ID is set on `tx.networkID`.

Separately, the filter wasn't effectively working for non-built-in
networks either. It was allowing notifications to be triggered for all
custom networks regardless of the current selected network.

These problems date back to the original implementation of incoming
transaction notifications (see https://github.com/MetaMask/metamask-mobile/pull/447/files#r261268602)

The filter has been updated to use chain ID rather than network ID,
and it gets the chain ID from `tx.chainId` which is guaranteed to be
set for all incoming transactions.

This relates to https://github.com/MetaMask/mobile-planning/issues/1226